Add Release API from job-server
This API is a local replica of the Release API in job-server to create
a release. It copies/snapshots the files locally, and then proxies the
request through to job-server.

The local copy is named after the returned release id, so it can be
referenced for serving files from to the SPA later.

Use httpx as it plays better with fastapi than requests.