Add signed token authentication
The token is signed via itsdangerous with the BACKEND_TOKEN. It
includes:

 - url prefix allowed
 - user
 - expiry
 - scope (simplified form of permissions for this token)

We use itsdangerous for the signing, so as to avoid a footgunning. We
use the default hmac algorithm, but with sha256, and we use the hmac key
derivation to protect against weak secrets (which is not as good as
HKDF, but will do for now). For secrets we use the BACKEND_TOKEN as the
signing key, but we also use the BACKEND name as an additional salt.

This replaces the previous auth method, and we now can use the user from
the token, and do not need OS-User header any more. It also saves us
from passing the actual secret back and forth over HTTP.

In the process, I refactored the pydantic models into a `schema` module.
The idea was to begin factoring out the stuff that will be extracted
into a shared package. This module just depends on stdlib and pydantic.

I also put the signing code into a signing module, with the same goal.
It also encapsulates all the token signing and verifying logic, so it is
all in one place, as we do not want to get that wrong. This module also
depends on `itsdangerous`, but I plan to make that an optional
dependency in the shared package.

I also added a couple missing field to FileSchema, to match job-server's
model.