Add release file index and serving apis
I release there was a security issue where a valid token for one
workspace could be used to access any release from any workspace. To
fix, I moved the releases to be stored inside their workspace dir, and
use that as an access control.

In the process, I refactored the test machinery out into reusable
factories.