Return 401 if the auth token has expired, rather than 403
This will allow the SPA to reload and refresh the token, allowing us to
keep quite short token lifetimes, hopefully.

This was actually a bit tricky to figure out how to do in pydantic.
The solution was eventually to add `root_validator` that runs after all
the field validations have been done. If the only invalid field is
`expiry`, then we raise a different, non-ValueError exception, which
can then be caught with `except:` as normal.