Refactor release-api for consistant url
We now rewrite the Location header for creating a release to point to
hatch rather than forwarding the job-server's Location header.

The intention is for hatch clients (osrelease and the SPA) as the
endpoint for uploading files.

To facilitate this we change the upload api url. Rather than passing the
name of the file as a path parameter, we include it in the request body.