Refactor docker gubbins into docker/ subdir
This avoids polluting the root dir with lots of files. Had to adjust
some paths in the docker-compose.yaml file, but otherwise works fine.

Due to the polluting issue above, I had previously commented out the
Dockerfile blocks for installing from `dependencies.txt` and
`build-dependencies.txt`, as we didn't need them in this project. But
I've now added empty ones in, to aid future crew when they need to add
dependencies.

For the just commands, I use trick where we can defer to a justfile in
a subdir. So `docker-build` now invokes `just docker/build`.

This allows us to keep the docker build stuff isolated, which avoids
complicating the main `justfile`. It also means we can more easily
pull in updates to `docker/justfile` in future iterations of the
repo-template, without conflicting with the main `justfile`