Refactor some docker based on prod set up
The main change is that we hardcode a prod uid/gid. This allows us to
use bind mounting directories in prod

Also clean up some of the docker justfile commands

 * rename `docker-run` to `docker-serve`
 * add `docker-run ...` as a wrapper round invoking `docker run ...`
 * add `docker-exec ...` as a wrapper round invoking `docker exec ...`

Also clean up the naming of images in docker-compose to drop the
unnecessary and brittle `-prod` suffix for the prod image.