Remove index endpoint trailing slashes
We wanted to avoid lots of redirects in the logs by having job-server
serve URLs with trailing slashes but due to the way the furl library
produces URLs that made life just slightly more difficult (lots of
`f.url + "/"` which had to be explained).  So it seemed simpler to
remove the trailing slashes here.  Since the SPA isn't using these URLs
to construct further URLs this won't affect other parts of the stack.