Fix bad test config that was hiding errors
A while back, we separated JOB_SERVER_ENDPOINT (which had a /api/v2 path
prefix) and SPA_ORIGIN (which did not). But the tests sadly used an old
test value of JOB_SERVER_ENDPOINT by default that was did not include
the path prefix. Thus all the api calls added a prefix, and everything
worked in test.

This happened to go unoticed, as the TPP install of release-hatch had an
incorrectly configured JOB_SERVER_ENDPOINT w/o the prefix, so stuff
worked.

However, in test and emis backends, where the JOB_SERVER_ENDPOINT does
have a prefix, as it is shared with job-runner config. So release-hatch
ended up double prefixing things, and thus 404s abounded.

The solution is not to have a seperate test config, and test the default
value, which includes the prefix.