fix: decouple validating hostname from path.
Previously, release-hatch validated that the full url we were handling
matched the token url prefix.

This meant we could only support one valid host. This has proved
problematic, as this included the port. For various reasons, EMIS users
access via one port when viewing outputs (because SSH forwarding)
and another when releasing outputs via `osrelease`. Also, localhost is
common access for testing and other operations, so it would be nice to
support that as an option.

This PR decouples the url validation into a hostname check (w/o ports),
and a path check.

It also tightens up the validation of urls in the signed token, so a url
must be valid.