Add support to request a release via the SPA.
This modifies the release creation API to support an alternate payload
POSTed from the SPA, rather than the current minimal osrelease payload.

This alternate payload is the same json format as the data we sent to
the SPA, so should be easy to reflect back. It includes an extra field,
`metadata`, to be populated by the SPA. We also remove the `user` field
as it was not used.

This consolidates the json schema to a single playload type, the
FileList, which is used for both listing files and requesting files for
review, which is much simpler. As such, it has been renamed to be
reflect its more general purpose.

There are few places where we have to branch to handle the appropriate
payload. Once the release UI in SPA is fully launched, we should be able
to delete the osrelease method.